requirements.txt
asyncio==3.4.3
aiopokeapi==0.1.4
numpy==1.23.4